refactor: UI 컴포넌트 및 Modifier 구조 재설계 및 패키지 정리
- 여러 파일로 분산되어 있던 `bottomBorder`, `dropShadow`, `horizontalFadingEdge` 등의 그리기 관련 Modifier들을 `modifierextention/DrawModifiers.kt`로 통합
- `ModifierExt.kt`의 패키지 위치를 `modifierextention`으로 이동하고, `noRippleClickable` 확장 함수를 해당 파일로 통합 관리
- `AfternoteCircularCheckbox`의 위치를 `icon` 패키지에서 `button` 패키지로 이동 및 관련 임포트 경로 수정
- 사용되지 않는 `AfternoteActionButton` 및 `RecordQnAListItem` 컴포저블 제거
- 코드 가독성 향상을 위해 `DrawModifiers.kt` 내 각 기능별 구분 주석 추가 및 KDoc 정비
