--- core/ui/src/main/kotlin/com/afternote/core/ui/BottomBorder.kt
@@ -1,35 +1,0 @@
-package com.afternote.core.ui
-
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.drawBehind
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
-
-/**
- * 하단 보더를 그리는 Modifier 확장 함수
- *
- * @param color 보더 색상
- * @param width 보더 두께 (기본값: 1.dp)
- */
-fun Modifier.bottomBorder(
-    color: Color,
-    width: Dp = 1.dp,
-) = this.then(
-    if (width > 0.dp && color != Color.Transparent) {
-        Modifier.drawBehind {
-            val borderWidth = width.toPx()
-            val y = size.height - (borderWidth / 2f)
-
-            drawLine(
-                color = color,
-                start = Offset(0f, y),
-                end = Offset(size.width, y),
-                strokeWidth = borderWidth,
-            )
-        }
-    } else {
-        Modifier
-    },
-)
